requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
Django==4.1.4
django-environ==0.9.0
django-pandas==0.6.6
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.38.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
ipykernel==6.20.1
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==5.1.3
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
parso==0.8.3
PGPy==0.6.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
PyDrive2==1.15.0
Pygments==2.14.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.8.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.6
